requirements/constraints.txt
flask
kafka-python